+++ GuildWars2.ArenaNet/Model/MappedModel.cs
@@ -5,6 +5,6 @@ namespace GuildWars2.ArenaNet.Model
 {
     public class MappedModel
     {
-        public List<int> Coord { get; set; }
+        public List<double> Coord { get; set; }
     }
 }
